+++ src/frequenz/sdk/timeseries/ev_charger_pool/_ev_charger_pool.py
@@ -26,7 +26,6 @@ from ..formula_engine._formula_generators import (
     EVChargerCurrentFormula,
     EVChargerPowerFormula,
     FormulaGeneratorConfig,
-    FormulaType,
 )
 from ._set_current_bounds import BoundsSetter, ComponentCurrentLimit
 from ._state_tracker import EVChargerState, StateTracker
@@ -181,61 +180,6 @@ class EVChargerPool:
             EVChargerPowerFormula,
             FormulaGeneratorConfig(
                 component_ids=self._component_ids,
-                formula_type=FormulaType.PASSIVE_SIGN_CONVENTION,
-            ),
-        )
-        assert isinstance(engine, FormulaEngine)
-        return engine
-
-    @property
-    def production_power(self) -> FormulaEngine[Power]:
-        """Fetch the total power produced by the EV Chargers in the pool.
-
-        This formula produces positive values when producing power and 0 otherwise.
-
-        If a formula engine to calculate EV Charger power is not already running, it
-        will be started.
-
-        A receiver from the formula engine can be created using the `new_receiver`
-        method.
-
-        Returns:
-            A FormulaEngine that will calculate and stream the production power of all
-                EV Chargers.
-        """
-        engine = self._formula_pool.from_power_formula_generator(
-            "ev_charger_production_power",
-            EVChargerPowerFormula,
-            FormulaGeneratorConfig(
-                component_ids=self._component_ids,
-                formula_type=FormulaType.PRODUCTION,
-            ),
-        )
-        assert isinstance(engine, FormulaEngine)
-        return engine
-
-    @property
-    def consumption_power(self) -> FormulaEngine[Power]:
-        """Fetch the total power consumed by the EV Chargers in the pool.
-
-        This formula produces positive values when consuming power and 0 otherwise.
-
-        If a formula engine to calculate EV Charger power is not already running, it
-        will be started.
-
-        A receiver from the formula engine can be created using the `new_receiver`
-        method.
-
-        Returns:
-            A FormulaEngine that will calculate and stream the consumption power of all
-                EV Chargers.
-        """
-        engine = self._formula_pool.from_power_formula_generator(
-            "ev_charger_consumption_power",
-            EVChargerPowerFormula,
-            FormulaGeneratorConfig(
-                component_ids=self._component_ids,
-                formula_type=FormulaType.CONSUMPTION,
             ),
         )
         assert isinstance(engine, FormulaEngine)